+++ src/components/instant-apps-measurement/test/instant-apps-measurement.e2e.ts
@@ -1,0 +1,11 @@
+import { newE2EPage } from '@stencil/core/testing';
+
+describe('instant-apps-measurement', () => {
+  it('renders', async () => {
+    const page = await newE2EPage();
+    await page.setContent('<instant-apps-measurement></instant-apps-measurement>');
+
+    const element = await page.find('instant-apps-measurement');
+    expect(element).toHaveClass('hydrated');
+  });
+});