+++ src/components/instant-apps-scoreboard/test/instant-apps-scoreboard.e2e.ts
@@ -1,0 +1,11 @@
+import { newE2EPage } from '@stencil/core/testing';
+
+describe('instant-apps-scoreboard', () => {
+  it('renders', async () => {
+    const page = await newE2EPage();
+    await page.setContent('<instant-apps-scoreboard></instant-apps-scoreboard>');
+
+    const element = await page.find('instant-apps-scoreboard');
+    expect(element).toHaveClass('hydrated');
+  });
+});