+++ src/interfaces/interfaces.d.ts
@@ -187,7 +187,7 @@ export interface ExtentSelector {
 }
 
 type ExpressionField = 'string' | 'number' | 'date' | 'coded-value' | 'range' | 'checkbox';
-
+type ActiveTool = 'distance' | 'point' | 'area' | 'clear';
 export type FilterLayer = __esri.FeatureLayer | __esri.ImageryLayer | __esri.WFSLayer | __esri.StreamLayer | __esri.CSVLayer | __esri.GeoJSONLayer;
 export type FilterQueryLayer = __esri.FeatureLayer | __esri.WFSLayer | __esri.CSVLayer | __esri.GeoJSONLayer;
 
@@ -221,3 +221,9 @@ export type PopoverPlacement =
   | 'trailing-end'
   | 'trailing'
   | 'trailing-start';
+
+export interface IMeasureConfiguration {
+  areaUnit?: __esri.AreaUnit;
+  linearUnit?: __esri.LengthUnit;
+  coordinateFormat?: string;
+}